+++ Client/scripts/globals.ts
@@ -42,7 +42,7 @@ function Debug(thing: any) {
     console.log(thing);
 }
 
-async function GetGridSize(url: string) {
+async function GetGridSize(url: string, customSize:number) {
     let grid = [0, 0];
     if (url.length) {
         grid = await new Promise((resolve) => {
@@ -51,11 +51,13 @@ async function GetGridSize(url: string) {
             tempImg.className = "temp-image";
             tempImg.addEventListener("load", () => {
                 const bounds = tempImg.getBoundingClientRect();
-                const width = Math.floor(bounds.width / 32);
-                const height = Math.floor(bounds.height / 32);
+                const width = Math.floor(bounds.width / customSize);
+                const height = Math.floor(bounds.height / customSize);
+                tempImg.remove();
                 resolve([width, height]);
             });
             tempImg.addEventListener("error", () => {
+                tempImg.remove();
                 resolve([0, 0]);
             });
             document.body.appendChild(tempImg);